requirements.txt
# Data Processing
pandas==2.1.4
numpy==1.24.3
openpyxl==3.1.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
kaleido==0.2.1

# Data Validation
pydantic==2.5.2
pandera==0.17.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging
loguru==0.7.2

# Type hints
typing-extensions==4.8.0

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
